refactor(frontend): split flat map/ into layers/, hooks/, geometry/ + delete dead ds/index.ts
55 files coexisted in a flat map/ directory mixing three distinct kinds:
React layer components, map-local hooks, and pure framework-free geometry
helpers. This makes the directory hard to scan (what's a component vs. a
utility?), and hides the clean geometry tier (adaptive-grid is shared by 9
siblings) that naturally operates with zero React or MapLibre imports.

Split by kind — one level deep, tests co-located:

  map/layers/    — 9 React/MapLibre layer+popover components
  map/hooks/     — 4 map-local hooks
  map/geometry/  — 13 pure .ts helpers

MapCanvas.tsx + MapCanvas.types.ts stay at the map/ root as the entry
surface. All import paths rewritten to use the @/ alias for stability.
External importers (App.tsx, data/bucket-aggregates.ts, e2e/state-artboard)
updated. No barrel files added.

Dead barrel `frontend/src/components/ds/index.ts` confirmed zero importers
(all consumers already used direct file paths) and deleted.

Verified: `npm run build -w @bird-watch/frontend` green (tsc -b + vite, 226
modules), `npm test -w @bird-watch/frontend` green (1549/1549 tests pass).
