--- frontend/src/components/ds/index.ts
@@ -1,21 +1,0 @@
-/**
- * Design-system primitives barrel export.
- * Phase 3–5 surfaces import from here.
- */
-export { StatusBlock } from './StatusBlock.js';
-export type { StatusBlockProps, StatusBlockState, StatusBlockSurface, StatusBlockTone } from './StatusBlock.js';
-
-export { Photo } from './Photo.js';
-export type { PhotoProps, PhotoLayout } from './Photo.js';
-
-export { FamilySilhouette } from './FamilySilhouette.js';
-export type { FamilySilhouetteProps, SilhouetteLayout } from './FamilySilhouette.js';
-
-export { ClusterPill } from './ClusterPill.js';
-export type { ClusterPillProps } from './ClusterPill.js';
-
-export { FilterSentence } from './FilterSentence.js';
-export type { FilterSentenceProps } from './FilterSentence.js';
-
-export { SortLabel } from './SortLabel.js';
-export type { SortLabelProps } from './SortLabel.js';
