+++ frontend/src/components/map/MapCanvas.types.ts
@@ -12,7 +12,7 @@
 import type { MultiPolygon } from 'geojson';
 import type { AggregatedBucket, FamilySilhouette, Observation } from '@bird-watch/shared-types';
 import type { SpeciesDictionary } from '../../data/use-species-dictionary.js';
-import type { AdaptiveTile, ResolvedGrid } from './adaptive-grid.js';
+import type { AdaptiveTile, ResolvedGrid } from './geometry/adaptive-grid.js';
 
 /**
  * Resolved per-cluster adaptive data — the unit the Concern B cache stores
