+++ frontend/src/components/map/geometry/adaptive-grid.test.ts
