+++ frontend/src/components/map/geometry/basemap-label-contrast.ts
