+++ frontend/src/components/map/geometry/basemap-null-filter.ts
