+++ frontend/src/components/map/geometry/basemap-style.ts
