+++ frontend/src/components/map/geometry/camera-config.test.ts
