+++ frontend/src/components/map/geometry/camera-config.ts
