+++ frontend/src/components/map/geometry/mask.test.ts
