+++ frontend/src/components/map/geometry/mask.ts
