+++ frontend/src/components/map/geometry/reconcile-viewport.ts
