+++ frontend/src/components/map/geometry/silhouette-fallback.test.ts
