+++ frontend/src/components/map/geometry/silhouette-fallback.ts
