+++ frontend/src/components/map/geometry/silhouette-sprite.test.ts
