+++ frontend/src/components/map/geometry/silhouette-sprite.ts
