+++ frontend/src/components/map/hooks/use-map-resize.ts
