+++ frontend/src/components/map/hooks/use-silhouette-catalogue.test.ts
