+++ frontend/src/components/map/layers/CellHoverPreview.tsx
@@ -1,8 +1,8 @@
 import type { CSSProperties } from 'react';
 import { createPortal } from 'react-dom';
-import type { SpeciesAggregate } from './adaptive-grid.js';
-import { prettyFamily } from '../../derived.js';
-import { formatCount } from '../../lib/format-count.js';
+import type { SpeciesAggregate } from '@/components/map/geometry/adaptive-grid.js';
+import { prettyFamily } from '../../../derived.js';
+import { formatCount } from '../../../lib/format-count.js';
 
 /**
  * `<CellHoverPreview>` — compact hover preview for an adaptive-grid cell
