+++ frontend/src/components/map/layers/CellPopover.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useId, useLayoutEffect, useRef, useState } from 'react';
 import type { CSSProperties, KeyboardEvent } from 'react';
 import { createPortal } from 'react-dom';
-import type { SpeciesAggregate } from './adaptive-grid.js';
-import { prettyFamily } from '../../derived.js';
-import { formatCount } from '../../lib/format-count.js';
+import type { SpeciesAggregate } from '@/components/map/geometry/adaptive-grid.js';
+import { prettyFamily } from '../../../derived.js';
+import { formatCount } from '../../../lib/format-count.js';
 
 /**
  * `<CellPopover>` — full popover for one family within an adaptive-grid cell
