+++ frontend/src/components/map/layers/ClusterListPopover.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useId, useRef, useState } from 'react';
 import type { KeyboardEvent } from 'react';
 import { createPortal } from 'react-dom';
-import type { FamilyAggregate, SpeciesAggregate } from './adaptive-grid.js';
-import { prettyFamily } from '../../derived.js';
-import { formatCount } from '../../lib/format-count.js';
+import type { FamilyAggregate, SpeciesAggregate } from '@/components/map/geometry/adaptive-grid.js';
+import { prettyFamily } from '../../../derived.js';
+import { formatCount } from '../../../lib/format-count.js';
 
 /**
  * `<ClusterListPopover>` — mobile / coarse-pointer sheet-style popover for
