+++ frontend/src/components/map/layers/MapMarkerHitLayer.test.tsx
