+++ frontend/src/components/map/layers/PresentationMarker.tsx
